giving gcds components their own namespace
