+++ GCDS.NetTemplate.MVC.Sanity/Controllers/HomeController.cs
@@ -1,4 +1,5 @@
-using GCDS.NetTemplate.Components;
+using GCDS.NetTemplate.Components.Gcds;
+using GCDS.NetTemplate.Components.Custom;
 using GCDS.NetTemplate.Models;
 using GCDS.NetTemplate.MVC.Sanity.Models;
 using GCDS.NetTemplate.Utils;
