--- GCDS.NetTemplate/Components/Breadcrumbs.cs
@@ -1,8 +1,0 @@
-﻿namespace GCDS.NetTemplate.Components
-{
-    public class Breadcrumbs : Common
-    {
-        public bool HideCanadaLink { get; set; } = false;
-        public List<Link>? Items { get; set; }
-    }
-}
