+++ GCDS.NetTemplate/Components/Gcds/Header.cs
@@ -1,7 +1,17 @@
-﻿namespace GCDS.NetTemplate.Components
+﻿namespace GCDS.NetTemplate.Components.Gcds
 {
+    /// <summary>
+    /// Class holder for the header properties as defined in the GCDS template
+    /// https://design-system.alpha.canada.ca/en/components/header/
+    /// </summary>
     public class Header : Common
     {
+        public enum SignatureVariantType
+        {
+            colour,
+            white
+        }
+
         /// <summary>
         /// Language toogle link
         /// Will be set inside the layout from the Template Model
@@ -15,34 +25,28 @@
         public string? SkipToHref { get; set; }
 
         /// <summary>
-        /// Remove the link to Canada.ca on the brand logo
+        /// Remove the link to the Canada.ca homepage from the signature
         /// </summary>
         public bool SignatureHasLink { get; set; } = true;
 
-        public enum SignatureVariantType
-        {
-            colour,
-            white
-        }
         /// <summary>
-        /// Choose a white or colour signature (brand logo)
-        /// colour by default
+        /// Switch the logo between color or all white
         /// </summary>
         public SignatureVariantType SignatureVariant { get; set; } = SignatureVariantType.colour;
 
         /// <summary>
         /// NOT IMPLEMENTED
         /// </summary>
+        [Obsolete("This slot is not implemented yet.")]
         public string? Banner { get; set; }
 
         /// <summary>
-        /// Add breadcrumbs bellow the header
-        /// Canada.ca will be added to the breadcrumbs unless removed
+        /// Slot for the breadcrumbs
         /// </summary>
         public Breadcrumbs? Breadcrumb { get; set; }
 
         /// <summary>
-        /// Add a Menu to the header
+        /// Slot for the Menu
         /// Either use the TopicMenu, the default menu that is generated by the template
         /// OR use the TopNav menu and create your own
         /// </summary>
@@ -55,13 +59,15 @@
         public Search? Search { get; set; }
 
         /// <summary>
-        /// NOT VALIDATED
+        /// NOT IMPLEMENTED
         /// </summary>
+        [Obsolete("This slot is not implemented yet.")]
         public string? SkipToNav { get; set; }
 
         /// <summary>
-        /// NOT VALIDATED
+        /// NOT IMPLEMENTED
         /// </summary>
+        [Obsolete("This slot is not implemented yet.")]
         public string? LangToggle { get; set; }
     }
 }
