+++ GCDS.NetTemplate/Components/ILink.cs
@@ -10,6 +10,11 @@ namespace GCDS.NetTemplate.Components
 
     public static class ListLinkExtensions
     {
+        /// <summary>
+        /// Creates a JSON string of the links
+        /// </summary>
+        /// <param name="links">links</param>
+        /// <returns>JSON string</returns>
         public static string ToJson(this IEnumerable<ILink> links)
         {
             var tranformedLinks = links.ToDictionary(link => link.Text, link => link.Href);
