+++ GCDS.NetTemplate/Components/IMenu.cs
@@ -1,5 +1,8 @@
 ﻿namespace GCDS.NetTemplate.Components
 {
+    /// <summary>
+    /// Enables interchanging object types for the Menu slot in the Header
+    /// </summary>
     public interface IMenu : ICommon
     {
     }
