--- GCDS.NetTemplate/Components/NavGroup.cs
@@ -1,8 +1,0 @@
-﻿namespace GCDS.NetTemplate.Components
-{
-    public class NavGroup : INavLink
-    {
-        public required string Label { get; set; }
-        public required List<ILink> Links { get; set; }
-    }
-}
