--- GCDS.NetTemplate/Components/TopicMenu.cs
@@ -1,7 +1,0 @@
-﻿namespace GCDS.NetTemplate.Components
-{
-    public class TopicMenu : Common, IMenu
-    {
-        public bool Home { get; set; }
-    }
-}
