+++ GCDS.NetTemplate/Views/Shared/Gcds/Breadcrumbs.cshtml
@@ -1,4 +1,4 @@
-﻿@using GCDS.NetTemplate.Components
+﻿@using GCDS.NetTemplate.Components.Gcds
 @model Breadcrumbs
 
 <gcds-breadcrumbs hide-canada-link="@Model.HideCanadaLink"
