+++ GCDS.NetTemplate/Views/Shared/Gcds/Footer.cshtml
@@ -1,4 +1,5 @@
 ﻿@using GCDS.NetTemplate.Components
+@using GCDS.NetTemplate.Components.Gcds
 @model Footer
 
 <gcds-footer lang="@Model.Lang"
