+++ GCDS.NetTemplate/Views/Shared/Gcds/Link.cshtml
@@ -1,4 +1,4 @@
-@using GCDS.NetTemplate.Components
+@using GCDS.NetTemplate.Components.Gcds
 @model Link
 
 <gcds-link href="@Model.Href"
