+++ GCDS.NetTemplate/Views/Shared/Gcds/NavLink.cshtml
@@ -1,4 +1,4 @@
-﻿@using GCDS.NetTemplate.Components
+﻿@using GCDS.NetTemplate.Components.Gcds
 @model Link
 
 <gcds-nav-link href="@Model.Href" slot="@Model.Slot">
