+++ GCDS.NetTemplate/Views/Shared/Gcds/Search.cshtml
@@ -1,4 +1,4 @@
-﻿@using GCDS.NetTemplate.Components
+﻿@using GCDS.NetTemplate.Components.Gcds
 @model Search
 
 <gcds-search slot="@Model.Slot"
