+++ GCDS.NetTemplate/Views/Shared/Gcds/TopicMenu.cshtml
@@ -1,4 +1,4 @@
-﻿@using GCDS.NetTemplate.Components
+﻿@using GCDS.NetTemplate.Components.Gcds
 @model TopicMenu
 
 <gcds-topic-menu slot="@Model.Slot"
