+++ GCDS.NetTemplate/Views/Shared/_Layout.Internal.cshtml
@@ -1,6 +1,6 @@
 ﻿@using GCDS.NetTemplate.Models
 @using GCDS.NetTemplate.Utils
-@using GCDS.NetTemplate.Components
+@using GCDS.NetTemplate.Components.Gcds
 @{
 	ITemplateModel webTemplateModel = (ITemplateModel)ViewData[Constants.TEMPLATE_MODEL];
 	if (webTemplateModel?.InternalAppHeader == null)
@@ -12,7 +12,7 @@
 <head>
 	<meta charset="utf-8" />
 	<meta name="viewport"
-	content="width=device-width, initial-scale=1.0" />
+		  content="width=device-width, initial-scale=1.0" />
 
 	<title>@webTemplateModel.PageTitle</title>
 
@@ -20,7 +20,7 @@
 
 	<style>
 		.hidden-link {
-			position:absolute;
+			position: absolute;
 			top: var(--gcds-header-skiptonav-top);
 			width: 100%;
 			text-align: center;
@@ -33,43 +33,43 @@
 				width: inherit;
 				z-index: 3;
 			}
-		.hidden-link gcds-link:not(:focus) {
-			height: 0;
-			width: 0;
-			overflow: hidden;
-			clip: rect(0,0,0,0);
-		}
+
+				.hidden-link gcds-link:not(:focus) {
+					height: 0;
+					width: 0;
+					overflow: hidden;
+					clip: rect(0,0,0,0);
+				}
 	</style>
 </head>
 <body>
-	<header
-		style="margin: var(--gcds-header-margin) !important">
+	<header style="margin: var(--gcds-header-margin) !important">
 		<nav airia-label="@webTemplateModel.InternalAppHeader.SkipTo.AiriaLabel"
-			class="hidden-link">
+			 class="hidden-link">
 			@{
 				var link = new Link
-			{
-				Href = "#" + Constants.SKIP_TO_CONTENT_ID,
-				Text = @webTemplateModel.InternalAppHeader.SkipTo.Text
-			};
+				{
+					Href = "#" + Constants.SKIP_TO_CONTENT_ID,
+					Text = @webTemplateModel.InternalAppHeader.SkipTo.Text
+				};
 			}
-			@await Html.PartialAsync("Link", link)
+			@await Html.PartialAsync("Gcds/Link", link)
 		</nav>
 
 		<gcds-container size="xl"
-		centered
-		main-container
-		style="@webTemplateModel.InternalAppHeader.StyleOverride">
+						centered
+						main-container
+						style="@webTemplateModel.InternalAppHeader.StyleOverride">
 			<gcds-grid columns="2fr 1fr"
-			align-items="center"
-			place-content="space-between">
+					   align-items="center"
+					   place-content="space-between">
 
 				<gcds-nav-link href="@webTemplateModel.InternalAppHeader.SiteTitle.Href">
 					@webTemplateModel.InternalAppHeader.SiteTitle.Text
 				</gcds-nav-link>
 
 				<gcds-lang-toggle href="@webTemplateModel.LanguageToggleHref"
-				style="text-align:end">
+								  style="text-align:end">
 				</gcds-lang-toggle>
 
 			</gcds-grid>
@@ -77,7 +77,7 @@
 
 		@if (webTemplateModel.Header?.Menu != null)
 		{
-			@await Html.PartialAsync("TopNav", webTemplateModel.Header.Menu)
+			@await Html.PartialAsync("Gcds/TopNav", webTemplateModel.Header.Menu)
 		}
 	</header>
 
@@ -90,7 +90,7 @@
 		</main>
 	</gcds-container>
 
-	@await Html.PartialAsync("Footer", webTemplateModel.Footer)
+	@await Html.PartialAsync("Gcds/Footer", webTemplateModel.Footer)
 
 	@await RenderSectionAsync("Scripts", required: false)
 </body>
